drones/simulator_api/requirements.txt
Flask==3.0.2
kafka-python==2.0.2
numpy==1.26.4



